requirements.txt
openai-whisper
opencc
argparse
numpy
pyaudio
webrtcvad
halo
wave
pycryptodemo
requests
soundfile